feat(authenticate pages): users should be able to signup and login
- Consume the signup and login endpoints
- Protect pages with token
- Rearrange folder structre

[Delivers]
