+++ package.json
@@ -14,8 +14,12 @@
     "axios": "^0.18.0",
     "dotenv": "^6.2.0",
     "font-awesome": "^4.7.0",
+    "jsonwebtoken": "^8.4.0",
     "vue": "^2.5.17",
-    "vue-router": "^3.0.2"
+    "vue-axios": "^2.1.4",
+    "vue-router": "^3.0.2",
+    "vue-toast-notification": "0.0.1",
+    "vuex": "^3.0.1"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "^3.2.0",
