+++ src/components/BookCard/bookCard.scss
