--- src/components/Dashboard/NewReleases/NewReleases.vue
@@ -1,25 +1,0 @@
-<template>
-    <BookGrid
-        :books="bookData" gridTitle="new Releases" 
-        customTitleClass="dashboard-title"
-        @modalToOpen="handleModalEvent"
-    />
-</template>
-
-<script>
-import BookGrid from '../../Shared/BookGrid/BookGrid.vue';
-import bookData from '../../../data/mockData';
-import modalMixins from '../../../mixins/modal-mixins';
-
-export default {
-    name: 'NewReleases',
-    mixins: [modalMixins],
-    data() {
-        return { bookData }
-    },
-    components: { BookGrid }
-}
-</script>
-
-<style lang="scss" scoped>
-</style>
