+++ src/components/Modal/modal.scss
