--- src/components/Shared/Styles.scss
@@ -1,249 +1,0 @@
-body {
-    box-sizing: border-box;
-    font-family: "Open Sans", Arial, sans-serif;
-    font-size: 16px;
-  }
-
-a {
-  color: #70D179;
-}
-
-#app {
-    font-family: "Open Sans", Arial, sans-serif;
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-    text-align: center;
-    color: snow;
-    background: #3C3B3F;  /* fallback for old browsers */
-    background: -webkit-linear-gradient(to right, #605C3C, #3C3B3F);  /* Chrome 10-25, Safari 5.1-6 */
-    background: linear-gradient(to right, #605C3C, #3C3B3F); /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */
-}
-
-.bs-btn {
-    color: snow;
-    background-color: #70D179;
-    margin: 0 10px;
-    font-size: 16px;
-    font-weight: 600;
-    &:hover, &:visited, &:focus {
-        color: snow;
-        background-color: rgb(72, 170, 82);
-    }
-}
-
-.bs-btn-light {
-    color: #70D179;
-    background-color: transparent;
-    margin: 0 10px;
-    font-size: 16px;
-    border: 1px solid #70D179;
-    font-weight: 600;
-    &:hover, &:visited, &:focus {
-        color: snow;
-        background-color: rgb(72, 170, 82);
-        border: none
-    }
-}
-
-.bs-box-shadow {
-    // border: 1px solid black;
-    box-shadow: 2px 0px 2px 2px red;
-}
-
-.bs-input {
-    height: auto;
-    
-}
-
-.inp {
-  position: relative;
-  margin: auto;
-  width: 100%;
-  max-width: 280px;
-  //transform: rotateX(45deg) rotateY(-10deg) rotate(25deg) scale(2)
-  .label {
-    position: absolute;
-    top: 16px;
-    left: 0;
-    font-size: 16px;
-    color: #9098A9;
-    font-weight: 500;
-    transform-origin: 0 0;
-    transition: all .2s ease;
-  }
-  .border {
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    height: 2px;
-    width: 100%;
-    background: #0077FF;
-    transform: scaleX(0);
-    transform-origin: 0 0;
-    transition: all .15s ease;
-  }
-  input {
-    -webkit-appearance: none;
-    width: 100%;
-    border: 0;
-    font-family: inherit;
-    padding: 12px 0;
-    height: 48px;
-    font-size: 16px;
-    font-weight: 500;
-    border-bottom: 2px solid #C8CCD4;
-    background: none;
-    border-radius: 0;
-    color: #223254;
-    transition: all .15s ease;
-    &:hover {
-        background: rgba(#223254,.03);
-    }
-    &:not(:placeholder-shown)
-      + span {
-        color: #5A667F;
-        transform: translateY(-26px) scale(.75);
-      }
-    &:focus {
-        background: none;
-        outline: none;
-        + span {
-            color: #0077FF;
-            transform: translateY(-26px) scale(.75);
-            + .border {
-                transform: scaleX(1);
-            }
-        }
-    }
-  }
-}
-
-.inp {
-    position: relative;
-    margin: auto;
-    width: 100%;
-    max-width: 280px;
-  }
-  .inp .label {
-    position: absolute;
-    top: 16px;
-    left: 0;
-    font-size: 16px;
-    color: #9098a9;
-    font-weight: 500;
-    transform-origin: 0 0;
-    transition: all 0.2s ease;
-  }
-  .inp .border {
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    height: 2px;
-    width: 100%;
-    background: #70D179;
-    transform: scaleX(0);
-    transform-origin: 0 0;
-    transition: all 0.15s ease;
-  }
-  .inp input {
-    -webkit-appearance: none;
-    width: 100%;
-    border: 0;
-    font-family: inherit;
-    padding: 12px 0;
-    height: 48px;
-    font-size: 16px;
-    font-weight: 500;
-    border-bottom: 2px solid #c8ccd4;
-    background: none;
-    border-radius: 0;
-    color: #223254;
-    transition: all 0.15s ease;
-  }
-  .inp input:hover {
-    background: rgba(34,50,84,0.03);
-  }
-  .inp input:not(:placeholder-shown) + span {
-    color: #5a667f;
-    transform: translateY(-26px) scale(0.75);
-  }
-  .inp input:focus {
-    background: none;
-    outline: none;
-    height: 30px;
-    margin: 0 auto;
-  }
-  .inp input:focus + span {
-    color: #70D179;
-    transform: translateY(-26px) scale(0.75);
-  }
-  .inp input:focus + span + .border {
-    transform: scaleX(1);
-  }
-  
-
-.books-container {
-  width: 100%;
-  background: #0077FF;
-}
-  
-::-webkit-scrollbar {
-  display: none;
-}
-
-.text-ellipsis {
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.bs-title {
-  text-align: left;
-  width: 80%;
-  margin: auto;
-  height: auto;
-  padding-top: 30px;
-  text-transform: capitalize;
-}
-
-input:focus {
-  border-bottom: 1px solid #70D179 !important;
-  -webkit-box-shadow: 0 1px 0 0 #70D179 !important;
-  box-shadow: 0 1px 0 0 #70D179 !important;
-  + label {
-    color: #70D179 !important;
-  }
-}
-
-
-
-.default-color, .bs-color {
-  color: #70D179 !important;
-}
-
-.bs-color-deep {
-  color: #48aa52 !important;
-}
-
-.bs-color-bg {
-  background-color: #70D179 !important;
-}
-
-.bs-color-bg-deep {
-  background-color: #48aa52 !important;
-}
-
-.bs-float-btn {
-  margin-bottom: 0px;
-  z-index: 5;
-}
-
-.bs-progress {
-  margin: 10px;
-  width: 97.5%;
-}
-
-.danger {
-  color: rgb(243, 114, 114);
-  font-size: 12px;
-}
