+++ src/components/Spinner/spinner.scss
@@ -1,0 +1,7 @@
+.bs-spinner {
+    // border: 1px solid blue;
+    margin: 10px 0 0 10px;
+    .bs-color {
+        border-color: snow 
+    }
+};
