+++ src/routes/index.js
@@ -1,14 +1,14 @@
 import Vue from 'vue';
 import Router from 'vue-router';
-import HomePage from '../components/Home/Home.vue';
-import Dash_board from '../components/Dashboard';
-import Dashboard from '../components/Dashboard/Dashboard.vue';
-import Favourite from '../components/Dashboard/Favourite/Favourite.vue';
-import BestSellers from '../components/Dashboard/BestSellers/BestSellers.vue';
-import NewReleases from '../components/Dashboard/NewReleases/NewReleases.vue';
-import AuthPage from '../components/AuthPage/AuthPage.vue';
-import Shelf from '../components/Shelf/Shelf.vue';
-import Profile from '../components/Profile/Profile.vue';
+import HomePage from '../views/Home/Home.vue';
+import Dash_board from '../views/Dashboard';
+import Dashboard from '../views/Dashboard/Dashboard.vue';
+import Favourite from '../views/Dashboard/Favourite/Favourite.vue';
+import BestSellers from '../views/Dashboard/BestSellers/BestSellers.vue';
+import NewReleases from '../views/Dashboard/NewReleases/NewReleases.vue';
+import AuthPage from '../views/AuthPage/AuthPage.vue';
+import Shelf from '../views/Shelf/Shelf.vue';
+import Profile from '../views/Profile/Profile.vue';
 
 Vue.use(Router);
 
@@ -64,8 +64,8 @@ export default new Router({
         {
             path: '/profile',
             name: 'Profile',
-            component: Profile
+            component: Profile,
         },
     ]
-})
+});
